Split up sam's library to isolate it better. Still works.
