+++ prj.conf
@@ -41,6 +41,7 @@ CONFIG_HEAP_MEM_POOL_SIZE=2048
 
 # Sam's extras
 CONFIG_I2C=y
+CONFIG_IMU_LOG_LEVEL_DBG=y
 
 # Enable channel management
 CONFIG_OPENTHREAD_CHANNEL_MONITOR=y
